+++ src/components/ExploreBtn.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
-import { SvgIcon, SvgWrapper } from '../assets/svg/SvgStyle';
-import SearchSvgIcon from '@/assets/svg/search.svg';
+import SearchSvgIcon from 'public/assets/svg/search.svg';
+import { SvgIcon, SvgWrapper } from './SvgStyle';
 
 const ExploreBtn = () => {
   return (
