+++ src/components/UserBtn.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
-import { SvgIcon, SvgWrapper } from '../assets/svg/SvgStyle';
-import UserSvgIcon from '@/assets/svg/user.svg';
+import UserSvgIcon from 'public/assets/svg/user.svg';
+import { SvgIcon, SvgWrapper } from './SvgStyle';
 
 const UserBtn = () => {
   return (
